python-ai/requirements_training.txt
# Training requirements for high-accuracy emotion detection model

# Core deep learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data processing
numpy>=1.24.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0

# Model export
onnx>=1.14.0
onnxruntime>=1.15.0
tf2onnx>=1.14.0
tensorflow>=2.13.0

# Utilities
tqdm>=4.65.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: For better performance
# tensorboard>=2.13.0  # For training visualization
# wandb>=0.15.0  # For experiment tracking

